Added Arith to StableHLO pass as an option and move to seperate file.
