+++ lib/Conversion/StableHLOToTTIR/CMakeLists.txt
@@ -6,6 +6,7 @@ include_directories(${PROJECT_SOURCE_DIR}/include)
 add_mlir_library(TTMLIRStableHLOToTTIR
   StableHLOToTTIRPatterns.cpp
   StableHLOToTTIRPass.cpp
+  ArithToStableHLOPass.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${PROJECT_SOURCE_DIR}/include/ttmlir/Conversion/StableHLOToTTIR
