+++ lib/Conversion/StableHLOToTTIR/StableHLOToTTIRPass.cpp
@@ -26,7 +26,6 @@ using namespace mlir::tt;
 namespace mlir::tt::ttir {
 
 #define GEN_PASS_DEF_CONVERTSTABLEHLOTOTTIR
-#define GEN_PASS_DEF_CONVERTARITHTOSTABLEHLO
 #include "ttmlir/Conversion/Passes.h.inc"
 
 } // namespace mlir::tt::ttir
@@ -107,49 +106,12 @@ struct ConvertStableHLOToTTIRPass
   }
 };
 
-struct ConvertArithToStableHLOPass
-    : public ttir::impl::ConvertArithToStableHLOBase<
-          ConvertArithToStableHLOPass> {
-  void runOnOperation() final {
-    mlir::ConversionTarget target(getContext());
-
-    target.addIllegalDialect<mlir::arith::ArithDialect>();
-    target.addLegalDialect<mlir::stablehlo::StablehloDialect>();
-    target.addLegalOp<mlir::ModuleOp>();
-    target.addLegalOp<mlir::func::FuncOp>();
-    target.addLegalOp<mlir::func::ReturnOp>();
-
-    // For now keep the same type assuming StableHLO ops operate on builtin
-    // tensor.
-    TypeConverter typeConverter;
-    typeConverter.addConversion([](Type type) {
-      assert(isa<RankedTensorType>(type) &&
-             "only ranked tensor type supported");
-      return type;
-    });
-    RewritePatternSet patterns(&getContext());
-
-    populateArithToStableHLOPatterns(&getContext(), patterns, typeConverter);
-
-    // Apply conversion.
-    if (failed(
-            applyFullConversion(getOperation(), target, std::move(patterns)))) {
-      signalPassFailure();
-      return;
-    }
-  }
-};
-
 } // namespace
 
 namespace mlir::tt {
 
 std::unique_ptr<OperationPass<ModuleOp>> createConvertStableHLOToTTIRPass() {
   return std::make_unique<ConvertStableHLOToTTIRPass>();
-}
-
-std::unique_ptr<OperationPass<ModuleOp>> createConvertArithToStableHLOPass() {
-  return std::make_unique<ConvertArithToStableHLOPass>();
 }
 
 } // namespace mlir::tt
